requirements.txt
django
django-filter
djangorestframework
djangorestframework-simplejwt
django-phonenumber-field
django-phonenumbers
django-simple-history
# sqlite
